Searching works when adding participants within conversation details. Conversation list searching updated for recent Atlas changes. Conversation keyboard disappearing fix when navigating from Details. Bump Atlas version

+++ Code/Controllers/ATLMApplicationViewController.m
@@ -196,7 +196,7 @@ static void *ATLMApplicationViewControllerObservationContext = &ATLMApplicationV
             break;
         }
         default:
-            [NSException raise:NSInternalInconsistencyException format:@"Unhandled ATLMApplicationState value=%lu", (NSUInteger)self.state];
+            [NSException raise:NSInternalInconsistencyException format:@"Unhandled ATLMApplicationState value=%lu", (unsigned long)self.state];
             break;
     }
 }
